:recycle:refactor: main.basic 파일에서 Header, ManulOverlay 화면 components 분리
