+++ src/components/Header.js
@@ -1,0 +1,11 @@
+// Header.js - 헤더 UI 컴포넌트 분리
+export function renderHeader() {
+  const header = document.createElement('div');
+  header.className = 'mb-8';
+  header.innerHTML = `
+    <h1 class="text-xs font-medium tracking-extra-wide uppercase mb-2">🛒 Hanghae Online Store</h1>
+    <div class="text-5xl tracking-tight leading-none">Shopping Cart</div>
+    <p id="item-count" class="text-sm text-gray-500 font-normal mt-3">🛍️ 0 items in cart</p>
+  `;
+  return header;
+}
